Further extracted functions and moved common functionality to the util.py module. Reconciled differences in the process_data functions and moved this logic to a single function
